Adds a worker for example
